Project structure simplification and refactoring
